refactor: Move config loading system into mopidy._app
