+++ tests/_app/test_default_config.py
@@ -1,0 +1,24 @@
+from mopidy import config
+from mopidy._app import config_schemas
+
+
+def test_core_schema_has_cache_dir():
+    assert "cache_dir" in config_schemas._core
+    assert isinstance(config_schemas._core["cache_dir"], config.Path)
+
+
+def test_core_schema_has_config_dir():
+    assert "config_dir" in config_schemas._core
+    assert isinstance(config_schemas._core["config_dir"], config.Path)
+
+
+def test_core_schema_has_data_dir():
+    assert "data_dir" in config_schemas._core
+    assert isinstance(config_schemas._core["data_dir"], config.Path)
+
+
+def test_core_schema_has_max_tracklist_length():
+    assert "max_tracklist_length" in config_schemas._core
+    max_tracklist_length_schema = config_schemas._core["max_tracklist_length"]
+    assert isinstance(max_tracklist_length_schema, config.Integer)
+    assert max_tracklist_length_schema._minimum == 1
